Resume write source lazily; suspend proactively
The write source is now only resumed when we know that data is
available, and is suspended as soon as it is all known to be written.

This resolves a crash arising due to a race condition, where `Writer`
relied on a no-op write event being scheduled in order to supsend the
source, but in the meantime a fast client sends more data to be written,
firing the read source and causing the write source to be double-resumed
before it had a chance to suspend itself.